requirements.txt
snowflake-snowpark-python
bs4
requests
